refactor: move user entity and user model to a shared module
