+++ src/modules/users/index.ts
@@ -1,2 +1,1 @@
-export * from './user.model'
 export * from './users.route'
