+++ src/modules/users/types/create-user-response.ts
@@ -1,4 +1,4 @@
-import {UserEntity} from '../user.entity'
+import {UserEntity} from '../../../shared/modules/user'
 
 export type CreateUserResponse = {
 	user: UserEntity
