+++ src/shared/modules/user/index.ts
@@ -1,0 +1,2 @@
+export * from './user.entity'
+export * from './user.model'
